My first patent application
When I was 15, I decided to make a portable keyboard to make it easier for visually impaired individuals to use electronic devices. This project was my first patent application. This invention not only enables visually impaired people to write on a virtual environment, but is compatible with every device, portable and practical compared to other keyboards. Thanks to this keyboard, the visually impaired individual types the letter, punctuation mark or number by typing the letter, punctuation mark or number according to the Braille alphabet. The text written in the Braille alphabet is written in the Latin alphabet on the virtual environment. This keyboard supports Turkish characters.
The attached files are in Turkish and are documents related to my patent application.