Spider robot with one servo motor
When I was in high school, I was in the computer and robotics club. One day, a club friend and I were discussing the minimum number of servo motors we could use to make a spider robot. I said I could do it with a single servo motor, but my friend said I couldn't. This was a little project I did to prove him wrong.